qik-test-qa-server/ai/requirements.txt
fastapi            # веб-фреймворк для нашего backend API
uvicorn            # сервер для запуска FastAPI
pydantic           # валидация входных/выходных данных
pydantic-settings  # настройки Pydantic из переменных окружения
pytest              # тестирование

allure-pytest

openai             # для выгрузки с работы с api LLM

langgraph          # создание агентных графов и оркестрация агентов
llama-index        # работа с контекстом, хранилищем данных, RAG-пайплайнами
llama-index-llms-openai  # драйвер для LLM-моделей в llama-index (можно заменить на другой)
llama-index-embeddings-openai  # генерация эмбеддингов (или заменить на локальные)

python-dotenv      # управление переменными окружения (.env)

httpx              # HTTP-клиент для запросов к внешним API (включая GitLab)
aiohttp            # асинхронный HTTP-клиент, если где-то понадобится

redis              # кэш, хранилище состояний агентов, очереди
celery             # фоновые задачи, если потребуется вынести тяжелые штуки
nats-py            # NATS asyncio клиент
gitpython          # Git утилиты (используется в CICD контроллере)

numpy              # базовая математика для обработки данных
pandas             # любые табличные данные (если агент будет анализировать что-то)
scikit-learn       # лёгкие модели ML, если какой-то агенту нужно что-то считать

pydub              # конвертация аудио (если понадобится при работе с голосом)
openai             # доступ к OpenAI API (если используем их модели)

loguru             # удобный логгер

langchain          # возможно для частей логики, которые проще собрать через LangChain

jinja2             # генерация шаблонов ответов и сообщений

python-multipart   # поддержка загрузки файлов в FastAPI (аудио, документы)

werkzeug           # утилиты для безопасности, работы с файлами и т.д.
